Windows install 6
Windows install 6

Windows install 6

Windows install 6

Windows install 6

Windows install 6

Update README.md

Windows install 7

Windows install 7

Windows install 7

+++ windows/utils.psm1
@@ -77,6 +77,9 @@ function Update-Firewall {
 
 Write-Output "Disable-TCC imported"
 function Disable-TCC {
+    If ($env:LOCAL  -eq 'yes')  {
+        return
+    }
     Write-Output "Disable TCC Mode on Nvidia Tesla GPU"
     $nvsmi = "C:\Program Files\NVIDIA Corporation\NVSMI\nvidia-smi.exe"
     $gpu = & $nvsmi --format=csv,noheader --query-gpu=pci.bus_id
@@ -161,6 +164,10 @@ function Install-FractalWallpaper ($run_on_cloud, $credentials) {
     # sleep for 15 seconds to make sure previous operations completed
     Start-Sleep -s 15
 
+    If ($env:LOCAL  -eq 'yes')  {
+        return
+    }
+
     # first download the wallpaper
     Write-Output "Downloading Fractal Wallpaper"
     $fractalwallpaper_name = "C:\Program Files\Fractal\Assets\wallpaper.png"
@@ -186,6 +193,9 @@ function Install-FractalWallpaper ($run_on_cloud, $credentials) {
 Write-Output "Disable-Cursor imported"
 function Disable-Cursor {
     # makes the Windows cursor blank to avoid duplicate cursor issue
+    If ($env:LOCAL  -eq 'yes')  {
+        return
+    }
     Write-Output "Downloading the Blank Cursor File"
     $cursorPath = "C:\Program Files\Fractal\Assets\blank.cur"
     $cursorPath_url = "https://fractal-cloud-setup-s3bucket.s3.amazonaws.com/blank.cur"
@@ -630,6 +640,10 @@ function Install-FractalServer ($protocol_branch) {
     $shared_libs_url = "https://fractal-protocol-shared-libs.s3.amazonaws.com/shared-libs.tar.gz"
     $webClient.DownloadFile($shared_libs_url, $shared_libs_name)
 
+    If ($env:LOCAL  -eq 'yes')  {
+        return
+    }
+
     Write-Output "Unzip the .tar.gz File and Remove shared-libs.tar.gz & /lib"
     tar -xvzf .\shared-libs.tar.gz
     Remove-Item -Path $shared_libs_name -Confirm:$false
@@ -710,6 +724,9 @@ function Install-NvidiaTeslaPublicDrivers {
 }
 
 function Set-OptimalGPUSettings {
+    If ($env:LOCAL  -eq 'yes')  {
+        return
+    }
     Write-Output "Setting Optimal Tesla M60 GPU Settings"
     C:\'Program Files'\'NVIDIA Corporation'\NVSMI\.\nvidia-smi --auto-boost-default=0
     C:\'Program Files'\'NVIDIA Corporation'\NVSMI\.\nvidia-smi -ac "2505,1177"
@@ -739,6 +756,9 @@ function Install-Unison {
 
 function Enable-SSHServer {
     Write-Output "Adding OpenSSH Server Capability"
+    If ($env:LOCAL  -eq 'yes')  {
+        return
+    }
     Add-WindowsCapability -Online -Name OpenSSH.Server~~~~0.0.1.0
     if (-not $?) {
         Write-Output "Add-WindowsCapability Failed, Trying DISM"
